% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19TargetAndOutcomeCharacterization.R
\docType{package}
\name{Covid19TargetAndOutcomeCharacterization-package}
\alias{Covid19TargetAndOutcomeCharacterization}
\alias{Covid19TargetAndOutcomeCharacterization-package}
\title{Covid19TargetAndOutcomeCharacterization: A Central Package for Evaluation the Cohorts and Outcomes amongst COVID-19 Studies}
\description{
A package for characterizing COVID-19 patients with and without various outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{}
}

}
\keyword{internal}
